Refactor subclasses to own cs file
